Update dns tool to take the stack name.
