+++ util/vpc-tools/vpc-tools.py
@@ -12,6 +12,7 @@ Options:
 """
 import boto
 from docopt import docopt
+from vpcutil import vpc_for_stack_name
 
 
 VERSION="vpc tools 0.1"
@@ -43,13 +44,6 @@ def dispatch(args):
     if args.get("ssh-config"):
         _ssh_config(args)
 
-def vpc_for_stack_name(stack_name):
-    cfn = boto.connect_cloudformation()
-    resources = cfn.list_stack_resources(stack_name)
-    for resource in resources:
-      if resource.resource_type == 'AWS::EC2::VPC':
-        return resource.physical_resource_id
-
 def _ssh_config(args):
     if args.get("vpc"):
       vpc_id = args.get("<vpc_id>")
@@ -57,7 +51,7 @@ def _ssh_config(args):
       stack_name = args.get("<stack_name>")
       vpc_id = vpc_for_stack_name(stack_name)
     else:
-      raise Exception("No way to know which vpc to query.")
+      raise Exception("No vpc_idao or stack_name provided.")
 
     vpc = boto.connect_vpc()
